separating demo into different files with clear purpose and cleanup
